requirements.txt
faiss-cpu
numpy
pymupdf
arabic-reshaper
python-bidi
requests
python-dotenv
huggingface-hub